// src/models/persona_model.ts
export class Persona{
    nombre: string;
    altura: string;
    peso: string;
    colorCabello: string;
    colorPiel: string;
    colorOjos: string;
    anhoNacimiento: string;
    genero: string;
    mundoNatal: string;
    peliculas: string[];
    especies: string[];
    vehiculos: string[];
    naves: string[];
    creado: string;
    editado: string;
    url: string;
}